Split out the err functions to a new file, no logic change
